Use uint64_t for maps and smaps for better support on 32-bit systems

+++ include/pfs/utils.hpp
@@ -165,7 +165,7 @@ ip parse_ipv6_address(const std::string& ip_address_hex);
 std::pair<ip, uint16_t> parse_address(const std::string& address_str);
 
 // Parses a memory size line (e.g. VmRSS:      4488 kB)
-void parse_memory_size(const std::string& value, size_t& out);
+void parse_memory_size(const std::string& value, uint64_t& out);
 
 } // namespace utils
 } // namespace impl
